implemented headerBehavior in HeaderConstraintDelegate. fixed build error
