--- PagingMenuWithHeader/HeaderWithPagingMenu/Base.lproj/Main.storyboard
@@ -1,82 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" standalone="no"?>
-<document type="com.apple.InterfaceBuilder3.CocoaTouch.Storyboard.XIB" version="3.0" toolsVersion="11542" systemVersion="15G1108" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" colorMatched="YES" initialViewController="BYZ-38-t0r">
-    <device id="retina4_7" orientation="portrait">
-        <adaptation id="fullscreen"/>
-    </device>
-    <dependencies>
-        <deployment identifier="iOS"/>
-        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="11524"/>
-        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
-    </dependencies>
-    <scenes>
-        <!--View Controller-->
-        <scene sceneID="tne-QT-ifu">
-            <objects>
-                <viewController id="BYZ-38-t0r" customClass="ViewController" customModule="PagingMenuWithHeader" customModuleProvider="target" sceneMemberID="viewController">
-                    <layoutGuides>
-                        <viewControllerLayoutGuide type="top" id="y3c-jy-aDJ"/>
-                        <viewControllerLayoutGuide type="bottom" id="wfy-db-euE"/>
-                    </layoutGuides>
-                    <view key="view" contentMode="scaleToFill" id="8bC-Xf-vdC">
-                        <rect key="frame" x="0.0" y="0.0" width="375" height="667"/>
-                        <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
-                        <subviews>
-                            <containerView opaque="NO" contentMode="scaleToFill" translatesAutoresizingMaskIntoConstraints="NO" id="AtF-DC-zYh">
-                                <rect key="frame" x="0.0" y="118" width="375" height="549"/>
-                                <connections>
-                                    <segue destination="d7C-Zc-2SA" kind="embed" id="bbs-Jk-9Vr"/>
-                                </connections>
-                            </containerView>
-                            <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="Header" textAlignment="center" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="RAi-45-OOM">
-                                <rect key="frame" x="0.0" y="-10" width="375" height="128"/>
-                                <color key="backgroundColor" red="1" green="0.80000000000000004" blue="0.73333333333333328" alpha="1" colorSpace="calibratedRGB"/>
-                                <constraints>
-                                    <constraint firstAttribute="height" constant="128" id="Eop-0d-I4J"/>
-                                </constraints>
-                                <fontDescription key="fontDescription" type="system" pointSize="20"/>
-                                <color key="textColor" red="1" green="1" blue="1" alpha="1" colorSpace="calibratedRGB"/>
-                                <nil key="highlightedColor"/>
-                            </label>
-                        </subviews>
-                        <color key="backgroundColor" red="1" green="1" blue="1" alpha="1" colorSpace="custom" customColorSpace="sRGB"/>
-                        <constraints>
-                            <constraint firstItem="AtF-DC-zYh" firstAttribute="leading" secondItem="8bC-Xf-vdC" secondAttribute="leading" id="Cit-z0-zrG"/>
-                            <constraint firstAttribute="trailing" secondItem="AtF-DC-zYh" secondAttribute="trailing" id="VcR-pD-TuB"/>
-                            <constraint firstItem="wfy-db-euE" firstAttribute="top" secondItem="AtF-DC-zYh" secondAttribute="bottom" id="c1B-VY-HIg"/>
-                            <constraint firstAttribute="trailing" secondItem="RAi-45-OOM" secondAttribute="trailing" id="hCk-pf-kTH"/>
-                            <constraint firstItem="AtF-DC-zYh" firstAttribute="top" secondItem="RAi-45-OOM" secondAttribute="bottom" id="ohe-tI-vqN"/>
-                            <constraint firstItem="RAi-45-OOM" firstAttribute="top" secondItem="y3c-jy-aDJ" secondAttribute="bottom" constant="-30" id="tGo-8w-GA8"/>
-                            <constraint firstItem="RAi-45-OOM" firstAttribute="leading" secondItem="8bC-Xf-vdC" secondAttribute="leading" id="vBT-2M-MP2"/>
-                        </constraints>
-                    </view>
-                    <connections>
-                        <outlet property="container" destination="AtF-DC-zYh" id="JlT-bH-CBF"/>
-                        <outlet property="headerConstraint" destination="tGo-8w-GA8" id="zJm-VQ-SJJ"/>
-                        <outlet property="headerView" destination="RAi-45-OOM" id="HwK-BD-LOo"/>
-                        <outlet property="profile" destination="RAi-45-OOM" id="7EM-Ks-UH5"/>
-                    </connections>
-                </viewController>
-                <placeholder placeholderIdentifier="IBFirstResponder" id="dkx-z0-nzr" sceneMemberID="firstResponder"/>
-            </objects>
-            <point key="canvasLocation" x="136.80000000000001" y="137.18140929535232"/>
-        </scene>
-        <!--Paging Menu Controller-->
-        <scene sceneID="WgQ-Ph-nS5">
-            <objects>
-                <viewController id="d7C-Zc-2SA" customClass="PagingMenuController" customModule="PagingMenuController" sceneMemberID="viewController">
-                    <layoutGuides>
-                        <viewControllerLayoutGuide type="top" id="2bn-5a-Aef"/>
-                        <viewControllerLayoutGuide type="bottom" id="ryN-Yd-yjd"/>
-                    </layoutGuides>
-                    <view key="view" contentMode="scaleToFill" id="vbX-My-HiM">
-                        <rect key="frame" x="0.0" y="0.0" width="375" height="549"/>
-                        <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
-                        <color key="backgroundColor" white="1" alpha="1" colorSpace="calibratedWhite"/>
-                    </view>
-                </viewController>
-                <placeholder placeholderIdentifier="IBFirstResponder" id="UGg-Wf-tXY" userLabel="First Responder" sceneMemberID="firstResponder"/>
-            </objects>
-            <point key="canvasLocation" x="929" y="137"/>
-        </scene>
-    </scenes>
-</document>
